Add formatter-only mode (no checkstyle)

+++ style-pre-commit-hook.sh
@@ -10,7 +10,7 @@
 # Configuration
 NO_FORMATTING_WHEN_CHECKSTYLE_OK=false # Set this to true to skip the (slow) IntelliJ autoformatter when there are no checkstyle violations on the changed files
 USE_AUTOFORMATTER=true # Set this to false to only run checkstyle and abort directly if violations are found (even if they could be fixed by running the autoformatter)
-USE_CHECKSTYLE=false # Set this to false to disable checkstyle and only apply the autoformatter
+USE_CHECKSTYLE=true # Set this to false to disable checkstyle and only apply the autoformatter
 ALLOW_IGNORE_CHECKSTYLE=true # Set this to true to allow committing even if this results in checking in checkstyle violations (warning will be shown)
 
 STYLE_REPO=$(dirname $0)
@@ -55,8 +55,8 @@ if [ "$USE_AUTOFORMATTER" = true ] && [ ! command -v $AUTOFORMATTER &> /dev/null
     exit 1
 fi
 
-if ! command -v checkstyle &> /dev/null; then
-    print_error "Codestyle pre-commit validation is enabled, but checkstyle could not be found. Please install it or disable this hook."
+if [ "$USE_CHECKSTYLE" = true ] && [ ! command -v checkstyle &> /dev/null ]; then
+    print_error "Codestyle pre-commit validation is enabled, but checkstyle could not be found. Please install checkstyle, set USE_CHECKSTYLE to false to disable checkstyle, or disable this hook completely."
     exit 1
 fi
 
@@ -110,7 +110,7 @@ fi
 
 # If corresponding setting enabled, skip autoformatter when no checkstyle violations are found
 CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=unknown
-if [ "$NO_FORMATTING_WHEN_CHECKSTYLE_OK" = true ]; then
+if [ "$USE_CHECKSTYLE" = true ] && [ "$NO_FORMATTING_WHEN_CHECKSTYLE_OK" = true ]; then
     print_info "Running checkstyle..."
     CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=false
     for FILE in $FILES; do
@@ -150,27 +150,29 @@ $AUTOFORMATTER -s $AUTOFORMATTER_CONFIG "${TEMP_COPIES[@]}" &> /dev/null
 # Early reject if there are checkstyle issues even after applying the autoformatter
 CHECKSTYLE_VIOLATION_FILES=()
 CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES=false
-[ -e "$CHECKSTYLE_LOG_FORMATTED" ] && rm "$CHECKSTYLE_LOG_FORMATTED"
-for FILE in $FILES; do
-    TEMP_FILE="$WORKING_COPY_DIR/$FILE"
-    print_info "File: $TEMP_FILE" >> "$CHECKSTYLE_LOG_FORMATTED"
-    checkstyle -c "$CHECKSTYLE_CONFIG" "$TEMP_FILE" &>> "$CHECKSTYLE_LOG_FORMATTED"
-    if [ $? -ne 0 ]; then
-        CHECKSTYLE_VIOLATION_FILES+=("$FILE")
-        CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES=true
-    fi
-    print_info "" >> "$CHECKSTYLE_LOG_FORMATTED"
-done
-if [ "$CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES" = true ]; then
-    print_info ""
-    print_error "Checkstyle violations found (with formatter applied) in the following file(s):"
-    for FILE in "${CHECKSTYLE_VIOLATION_FILES[@]}"; do
-        print_error "    $FILE"
+if [ "$USE_CHECKSTYLE" = true ]; then
+    [ -e "$CHECKSTYLE_LOG_FORMATTED" ] && rm "$CHECKSTYLE_LOG_FORMATTED"
+    for FILE in $FILES; do
+        TEMP_FILE="$WORKING_COPY_DIR/$FILE"
+        print_info "File: $TEMP_FILE" >> "$CHECKSTYLE_LOG_FORMATTED"
+        checkstyle -c "$CHECKSTYLE_CONFIG" "$TEMP_FILE" &>> "$CHECKSTYLE_LOG_FORMATTED"
+        if [ $? -ne 0 ]; then
+            CHECKSTYLE_VIOLATION_FILES+=("$FILE")
+            CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES=true
+        fi
+        print_info "" >> "$CHECKSTYLE_LOG_FORMATTED"
     done
-    print_info ""
-    if [ "$ALLOW_IGNORE_CHECKSTYLE" = false ]; then
-        print_error "Commit cancelled. Please fix the checkstyle issues."
-        exit 1
+    if [ "$CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES" = true ]; then
+        print_info ""
+        print_error "Checkstyle violations found (with formatter applied) in the following file(s):"
+        for FILE in "${CHECKSTYLE_VIOLATION_FILES[@]}"; do
+            print_error "    $FILE"
+        done
+        print_info ""
+        if [ "$ALLOW_IGNORE_CHECKSTYLE" = false ]; then
+            print_error "Commit cancelled. Please fix the checkstyle issues."
+            exit 1
+        fi
     fi
 fi
 
@@ -205,26 +207,33 @@ if [ ${#FILES_TO_FORMAT[@]} -eq 0 ]; then
     exit 1
 fi
 
+
 # Determine if there are checkstyle violations on the original files, requiring the user to accept the autoformatted changes
+
+# Performance improvement: Assume that formatter does not introduce checkstyle violations
+if [ "$CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES" = true ]; then
+    CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=true
+fi
+
+# If checkstyle disabled, proceed as if they are no violations
+if [ "$USE_CHECKSTYLE" = false ]; then
+    CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=false
+fi
+
 if [ "$CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES" = unknown ]; then
-    # Performance improvement: Assume that formatter does not introduce checkstyle violations
-    if [ "$CHECKSTYLE_VIOLATIONS_ON_FORMATTED_FILES" = true ]; then
-        CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=true
-    else
-        [ -e "$CHECKSTYLE_LOG_ORIGINAL" ] && rm "$CHECKSTYLE_LOG_ORIGINAL"
-        print_info "Running checkstyle on original files..."
-        CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=false
-        for FILE in $FILES; do
-            ORIGINAL_FILE=$(git rev-parse --show-toplevel)/$FILE
-            print_info "File: $ORIGINAL_FILE" >> "$CHECKSTYLE_LOG_ORIGINAL"
-            checkstyle -c "$CHECKSTYLE_CONFIG" "$ORIGINAL_FILE" &>> "$CHECKSTYLE_LOG_ORIGINAL"
-            if [ $? -ne 0 ]; then
-                CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=true
-                break # We can skip checking other files because the user will now definitly have to accept the autoformatter
-            fi
-            print_info "" >> "$CHECKSTYLE_LOG_ORIGINAL"
-        done
-    fi
+    [ -e "$CHECKSTYLE_LOG_ORIGINAL" ] && rm "$CHECKSTYLE_LOG_ORIGINAL"
+    print_info "Running checkstyle on original files..."
+    CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=false
+    for FILE in $FILES; do
+        ORIGINAL_FILE=$(git rev-parse --show-toplevel)/$FILE
+        print_info "File: $ORIGINAL_FILE" >> "$CHECKSTYLE_LOG_ORIGINAL"
+        checkstyle -c "$CHECKSTYLE_CONFIG" "$ORIGINAL_FILE" &>> "$CHECKSTYLE_LOG_ORIGINAL"
+        if [ $? -ne 0 ]; then
+            CHECKSTYLE_VIOLATIONS_ON_ORIGINAL_FILES=true
+            break # We can skip checking other files because the user will now definitly have to accept the autoformatter
+        fi
+        print_info "" >> "$CHECKSTYLE_LOG_ORIGINAL"
+    done
 fi
 
 # Prompt the user about the proposed changes
